[Refactor] Make current user button a component
